+++ p_unt.c
@@ -6,7 +6,7 @@
  *
  * Return: int
  */
-int print_unt(unsigned int num)
+int print_unt(unsigned long int num)
 {
 	unsigned int x = num;
 	int i = 0;
